Fix deleting multiple recordings with VDR-2.3.2
VDR-2.3.2 updates recording IDs while processing subsequent DELR
commands. This results in deleting wrong recordings. For example, if
LSTR returns IDs 1, 2, 3, 4 and
  delr 2
  delr 3
  quit
is issued, then recordings with IDs 2 and 4 (from the LSTR output) are
deleted. As a workaround, IDs are sorted in descending order before
sending a series of DELR commands.